requirements.txt
python-nmap>=0.7.1
streamlit>=1.28.0
requests>=2.31.0
